Adding Part 4 - Power function and Demo example
This is a first version of the functional to calculate power that should be ready to merge to master. It contains modifications suggested by Kristin in last pull request #17.  As always, the datasets used for debugging were various but limited. We will have to continue debugging, but this process may be more efficient once we start to use this in 'real' datasets and case studies. I hope this is a good start for that.

FYI: there may be enough good reasons to improve the way the logic of these functions works. Feel free to further develop, ideally try to continue following the standards of branching, communicating to the rest and asking for review - that will help improve the work I was able to do on this. From personal circumstances I will not be able to work on this for a while, but will try to help people use them and collect feedback for improvement / let people take the work further.

Juan